move orb_endpoints::orb_id to orb_info::orb_id
