+++ endpoints/Cargo.toml
@@ -13,3 +13,4 @@ rust-version.workspace = true
 [dependencies]
 hex = "0.4.3"
 url = "2.5.0"
+orb-info.workspace = true
