+++ fleet-cmdr/Cargo.toml
@@ -21,6 +21,7 @@ listenfd = "1.0.0"
 once_cell = "1.15.0"
 orb-build-info.workspace = true
 orb-endpoints.workspace = true
+orb-info.workspace = true
 orb-telemetry.workspace = true
 secrecy = { workspace = true, features = ["serde"] }
 serde = { workspace = true, features = ["derive"] }
